backend/requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.0
catboost==1.2.7
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.68.1
grpcio-status==1.68.1
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
lightgbm==4.5.0
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.4
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
rsa==4.9
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
